+++ packages/frontpage/app/(app)/_components/post-card.tsx
@@ -5,8 +5,9 @@ import { ensureUser, getUser } from "@/lib/data/user";
 import { TimeAgo } from "@/lib/components/time-ago";
 import { VoteButton } from "./vote-button";
 import { PostCollection } from "@/lib/data/atproto/post";
-import { DID, getVerifiedHandle } from "@/lib/data/atproto/did";
+import { getVerifiedHandle } from "@/lib/data/atproto/identity";
 import { UserHoverCard } from "@/lib/components/user-hover-card";
+import type { DID } from "@/lib/data/atproto/did";
 
 type PostProps = {
   id: number;
