+++ packages/frontpage/app/(app)/post/[postAuthor]/[postRkey]/layout.tsx
@@ -3,7 +3,7 @@ import { notFound } from "next/navigation";
 import { PostCard } from "../../../_components/post-card";
 import { DeletePostButton } from "./_lib/delete-post-button";
 import { getPost } from "@/lib/data/db/post";
-import { getDidFromHandleOrDid } from "@/lib/data/atproto/did";
+import { getDidFromHandleOrDid } from "@/lib/data/atproto/identity";
 
 type Params = {
   postRkey: string;
