+++ packages/frontpage/lib/db.ts
@@ -1,11 +1,8 @@
 import "server-only";
 
-import { loadEnvConfig } from "@next/env";
 import { drizzle } from "drizzle-orm/libsql";
 import { createClient } from "@libsql/client";
 import * as schema from "./schema";
-
-loadEnvConfig(process.cwd());
 
 const client = createClient({
   url: process.env.TURSO_CONNECTION_URL!,
